AudioRenderer fixes and additions
Fixed pitch factor + added looping and playing melodies.

+++ Code/AbstractSyntax/Melody.cs
@@ -2,7 +2,7 @@ namespace AbstractSyntax;
 
 public class Melody
 {
-	public int Length;
+	public float Length;
 	public string Id = "";
 	public List<string> SampleIds = new();
 	public List<Note> Notes = new();
